분류

+++ SpMVC_ReadBook/src/main/java/com/biz/book/mapper/BookDao.java
@@ -6,6 +6,6 @@ import com.biz.book.model.BookVO;
  * GenericDao 를 extends 하면 기본 CRUD method를
  * 별도로 정의하지 않아도 된다.
  */
-public interface BookDao extends GenericDao<BookVO, Integer> {
+public interface BookDao extends GenericDao<BookVO, Long> {
 
 }
